requirements.txt
speechrecognition==3.10.1
pyaudio==0.2.11
pynput==1.7.6
opencv-python==4.8.1.78
google-generativeai==0.3.2
Pillow==10.1.0 